feat: Status Effect Modifier function

+++ common/src/main/java/com/wynntils/utils/type/NamedValue.java
@@ -1,10 +1,10 @@
 /*
- * Copyright © Wynntils 2023.
+ * Copyright © Wynntils 2023-2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.utils.type;
 
-public record NamedValue(String name, int value) {
+public record NamedValue(String name, double value) {
     public static final NamedValue EMPTY = new NamedValue("", 0);
 
     @Override
